Moved components into relevant folders to try out a new organization style.
