+++ client/src/events/EventDetail.js
